ITK4: fork itkBinaryDilateImageFilter and itkBinaryErodeImageFilter
This classes have been gfork and static variables have been removed to avoid dl.close error in applications.
I used itykv4 version of classes to do the fork.
Forked classes are only use in the BinaryMorphological applications. It is still the internal itkv3
which is used in OTB.
Need to report this problem to ITK in JIRA, when the problem is solve on the ITK side we would safely remove these classes and use ITK classes

--HG--
extra : rebase_source : 2e1643bfda9ff314ef85bc2cf4303e6164032e63