BUG: corrections in ReflectanceToSurfaceReflectanceFilter class
Fistly, the spectral sentivity values tabulate in metadata class must be read by this filter and not by third part application. When no value is available and when there is no user spectral sentivity file provided, default sensitivity filters to 1 are used.
Secondly, the atmospheric coefficients compute by 6S were reordered by the filter using the BandIndexToWavelenghtPosition method. It is wrong in my opinion as it shows on Pleiades image where the red band is the most impacted by the atmosphere as It must be the blue band. I change this and consider that 6S returns atmospheric parameters in the same order as the band in the image file (note that spectral sensitivity tabulates in the metadata class are provided in the same order as bands in the image file.

--HG--
extra : rebase_source : 26c6b9f6c7b9b349e69a7cf50b3e7c90c6ca54a8