TEST: color mapping application tests are no longer dependents
Not that instead of using output of one application as input. I used Baseline data of the first application as input of the second one. It does not strickly break the dependency between the tests but avoid to push new images to the OTB-Data repository.

--HG--
extra : rebase_source : 0a2f1d78a1d5501891f71af0ffdbea67ba819375